requirements.txt
mysql-connector-python
tabulate
ttkbootstrap
customtkinter
streamlit
streamlit transformers
